move udf tests to test_udf_files
